issue #17 - fixed bad formatter Data Class Json

+++ core/src/main/kotlin/id/barakkastudio/core/data/model/Product.kt
@@ -17,7 +17,7 @@ data class Product(
     @SerializedName("images")
     val images: List<String?>? = null,
     @SerializedName("price")
-    val price: Long? = null,
+    val price: Double? = null,
     @SerializedName("rating")
     val rating: Double? = null,
     @SerializedName("stock")
